+++ src/Bridge/Ollama/CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =========
 
+0.4
+---
+
+ * [BC BREAK] The `hostUrl` parameter for `OllamaClient` has been removed
+ * [BC BREAK] The `host` parameter for `OllamaApiCatalog` has been removed
+ * [BC BREAK] The `hostUrl` parameter for `PlatformFactory::create()` has been renamed `endpoint`
+
 0.1
 ---